refactor: split lnv2 federation api extension trait
